api/requirements.txt
Flask==3.0.3
opencv-python==4.10.0.84
google-cloud-vision
